refactor(coding): relocate terminal presentation adapters
