+++ docs/internals/architecture/tui/README.md
@@ -29,12 +29,14 @@ Target source entrypoints remain:
 - `src/loushang/tui/`
 - `src/loushang/harnesstui/`
 - `src/loushang/coding/ui/`
+- `src/loushang/coding/presentation/tui/`
 
 `loushang.tui` is the generic terminal UI framework. The
 [`loushang.harnesstui`](../harnesstui/README.md) composition layer adapts neutral
-Harness conversation contracts into reusable TUI interaction. Product adapters
-such as `loushang.coding.ui` provide Coding-specific state, event projection,
-commands, policy, and runtime assembly.
+Harness conversation contracts into reusable TUI interaction.
+`loushang.coding.presentation.tui` owns Coding-specific raw event, tool,
+history, plain, and screen projection adapters. `loushang.coding.ui` retains
+concrete UI state, product surfaces, terminal bindings, and runtime composition.
 
 For status presentation, `loushang.tui` owns the generic status-bar widget and
 its layout, styling, invalidation, and rendering mechanics. A shared Harness
