+++ src/loushang/coding/presentation/tui/__init__.py
@@ -1,0 +1,1 @@
+"""Coding-owned terminal presentation adapters."""
