+++ src/loushang/coding/presentation/tui/plain.py
@@ -1,21 +1,104 @@
 from __future__ import annotations
 
+import shutil
 from dataclasses import dataclass, field
 from typing import Any, cast
 
-from loushang.coding.ui.conversation_event_adapter import (
-    CodingConversationEventAdapter,
+from loushang.coding.presentation.tui.events import CodingConversationEventAdapter
+from loushang.coding.presentation.tui.tool_transcript import (
+    ToolTranscriptProjector,
+    tool_block_to_record,
 )
-from loushang.coding.ui.plain_renderer import PlainCodingUiRenderer
-from loushang.coding.ui.tool_blocks import ToolTranscriptProjector
 from loushang.harness.presentation import ToolDefinitionResolver
 from loushang.harnesstui.conversation.plain_target import (
     PlainConversationProjectionTarget,
 )
 from loushang.harnesstui.conversation.projection import ConversationProjector
-from loushang.harnesstui.conversation.tool_transcript import (
-    ToolCallSnapshot,
+from loushang.harnesstui.conversation.tool_transcript import ToolCallSnapshot
+from loushang.harnesstui.plain.renderer import (
+    PlainConversationGlyphs,
+    PlainConversationProfile,
+    PlainConversationRenderer,
 )
+from loushang.tui.cell_width import strip_control_sequences
+from loushang.tui.render import MarkdownBlock as MarkdownBlock
+from loushang.tui.transcript import (
+    AssistantMessageRecord,
+    DisplayRecord,
+    ErrorRecord,
+    ToolExecutionRecord,
+    UserPromptRecord,
+    WorkedDividerRecord,
+)
+
+_INTERRUPTION_MESSAGE = (
+    "Conversation interrupted - tell the model what to do differently. "
+    "Something went wrong? Hit `/feedback` to report the issue."
+)
+
+
+def _coding_terminal_columns() -> int:
+    return shutil.get_terminal_size((80, 24)).columns
+
+
+def _coding_line(line: str, record: DisplayRecord) -> str:
+    line = strip_control_sequences(line)
+    if isinstance(record, UserPromptRecord) and line.startswith("> "):
+        return "› " + line[2:]
+    if isinstance(record, AssistantMessageRecord) and line.startswith("* "):
+        return "• " + line[2:]
+    if isinstance(record, ErrorRecord) and line.startswith("! Error: "):
+        return "■ Error: " + line[len("! Error: ") :]
+    if isinstance(record, ToolExecutionRecord):
+        if line.startswith("- Ran "):
+            return "• Ran " + line[len("- Ran ") :]
+        if line.startswith("! Ran "):
+            return "■ Ran " + line[len("! Ran ") :]
+    if isinstance(record, WorkedDividerRecord) and line.startswith("- Worked for "):
+        return line.replace("-", "─", 1).replace("-", "─")
+    return line
+
+
+_CODING_PLAIN_CONVERSATION_PROFILE = PlainConversationProfile(
+    title="Loushang TUI",
+    interruption_message=_INTERRUPTION_MESSAGE,
+    glyphs=PlainConversationGlyphs(
+        user_prompt="› ",
+        assistant="• ",
+        item="• ",
+        error="■ Error: ",
+        interruption="■ ",
+        rule="─",
+    ),
+    line_mapper=_coding_line,
+    tool_block_projector=tool_block_to_record,
+    terminal_columns=_coding_terminal_columns,
+)
+
+
+@dataclass
+class PlainCodingUiRenderer(PlainConversationRenderer):
+    """Coding presentation profile over the shared plain renderer."""
+
+    profile: PlainConversationProfile = field(
+        default=_CODING_PLAIN_CONVERSATION_PROFILE,
+        init=False,
+        repr=False,
+    )
+
+
+def extract_text(value: object) -> str:
+    content = getattr(value, "content", value)
+    if isinstance(content, str):
+        return content
+    if isinstance(content, list):
+        parts: list[str] = []
+        for part in content:
+            text = getattr(part, "text", None)
+            if isinstance(text, str):
+                parts.append(text)
+        return "".join(parts)
+    return ""
 
 
 @dataclass(init=False)
@@ -50,9 +133,7 @@ class PlainCodingEventRenderer:
             max_body_lines=max_tool_body_lines,
         )
         self._projection = ConversationProjector(
-            target=PlainConversationProjectionTarget(
-                renderer=renderer,
-            ),
+            target=PlainConversationProjectionTarget(renderer=renderer),
             tool_projector=self._tool_projector.neutral_projector,
             measure_tool_elapsed=False,
             tool_finish_cleanup="before_projection",
@@ -119,4 +200,8 @@ class PlainCodingEventRenderer:
         self._adapter.handle(event)
 
 
-__all__ = ["PlainCodingEventRenderer"]
+__all__ = [
+    "PlainCodingEventRenderer",
+    "PlainCodingUiRenderer",
+    "extract_text",
+]
