+++ tests/coding/test_screen_coding_tui_events.py
@@ -25,8 +25,8 @@ def _assistant(text: str = "", *, stop_reason: str = "stop", error_message: str 
 
 
 def test_screen_event_projector_streams_assistant_to_draft_then_commits_once() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     projector = ScreenCodingEventProjector(app)
@@ -53,8 +53,8 @@ def test_screen_event_projector_streams_assistant_to_draft_then_commits_once() -
 
 @pytest.mark.tui_render_contract
 def test_screen_event_projector_promotes_streaming_cache_through_shared_target() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
     from loushang.tui import RenderConstraints
 
     app = ScreenCodingTuiApp(
@@ -94,8 +94,8 @@ def test_screen_event_projector_promotes_streaming_cache_through_shared_target()
 
 
 def test_screen_event_projector_requires_assistant_message_for_delta() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     projector = ScreenCodingEventProjector(app)
@@ -112,8 +112,8 @@ def test_screen_event_projector_requires_assistant_message_for_delta() -> None:
 
 
 def test_screen_event_projector_renders_assistant_error_from_agent_end() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     projector = ScreenCodingEventProjector(app)
@@ -129,8 +129,8 @@ def test_screen_event_projector_renders_assistant_error_from_agent_end() -> None
 
 
 def test_screen_event_projector_commits_error_message_and_deduplicates_error() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     projector = ScreenCodingEventProjector(app)
@@ -151,8 +151,8 @@ def test_screen_event_projector_commits_error_message_and_deduplicates_error() -
 
 
 def test_screen_event_projector_commits_intentional_abort_without_error_record() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     projector = ScreenCodingEventProjector(app)
@@ -169,8 +169,8 @@ def test_screen_event_projector_commits_intentional_abort_without_error_record()
 
 
 def test_screen_event_projector_renders_user_message_and_skips_optimistic_echo() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     message = UserMessage(role="user", content=[TextPart(type="text", text="你好")], timestamp=0.0)
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
@@ -181,8 +181,8 @@ def test_screen_event_projector_renders_user_message_and_skips_optimistic_echo()
 
 
 def test_screen_event_projector_skips_user_message_when_matching_pending_echo() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     message = UserMessage(role="user", content=[TextPart(type="text", text="你好")], timestamp=0.0)
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
@@ -197,8 +197,8 @@ def test_screen_event_projector_skips_user_message_when_matching_pending_echo() 
 
 
 def test_screen_event_projector_drops_stale_pending_echo_on_mismatch() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     app.start_prompt("same", started_at=1.0)
@@ -225,8 +225,8 @@ def test_screen_event_projector_drops_stale_pending_echo_on_mismatch() -> None:
 
 
 def test_screen_event_projector_updates_tool_record_in_place() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 5.0)
     projector = ScreenCodingEventProjector(app, now=lambda: 5.0)
@@ -247,8 +247,8 @@ def test_screen_event_projector_updates_tool_record_in_place() -> None:
 
 
 def test_screen_event_projector_preserves_tool_elapsed_clock_boundaries() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(
         model_label="kimi",
@@ -287,8 +287,8 @@ def test_screen_event_projector_preserves_tool_elapsed_clock_boundaries() -> Non
 
 
 def test_screen_agent_start_does_not_read_clock_while_run_is_active() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(
         model_label="kimi",
@@ -311,8 +311,8 @@ def test_screen_agent_start_does_not_read_clock_while_run_is_active() -> None:
 
 
 def test_screen_event_projector_recovers_tool_update_without_start() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 5.0)
     projector = ScreenCodingEventProjector(app, now=lambda: 5.0)
@@ -332,8 +332,8 @@ def test_screen_event_projector_recovers_tool_update_without_start() -> None:
 
 
 def test_screen_event_projector_syncs_pending_queues() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     projector = ScreenCodingEventProjector(
@@ -349,8 +349,8 @@ def test_screen_event_projector_syncs_pending_queues() -> None:
 
 
 def test_screen_event_projector_preserves_coding_status_copy() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(
         model_label="kimi",
@@ -383,8 +383,8 @@ def test_screen_event_projector_preserves_coding_status_copy() -> None:
 
 
 def test_screen_event_projector_renders_queued_steer_into_transcript() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     app.start_prompt("初始问题", started_at=1.0)
@@ -411,8 +411,8 @@ def test_screen_event_projector_renders_queued_steer_into_transcript() -> None:
 
 
 def test_screen_event_projector_renders_queued_followup_into_transcript() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     app.start_prompt("初始问题", started_at=1.0)
@@ -439,8 +439,8 @@ def test_screen_event_projector_renders_queued_followup_into_transcript() -> Non
 
 
 def test_screen_event_projector_renders_same_text_queued_message_after_initial_echo() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     app.start_prompt("same", started_at=1.0)
@@ -454,8 +454,8 @@ def test_screen_event_projector_renders_same_text_queued_message_after_initial_e
 
 
 def test_screen_event_projector_appends_compaction_record_and_tracks_baseline_reset() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(model_label="kimi", cwd="/repo", branch="main", session_label="abcd", now=lambda: 1.0)
     app.state.records.extend(
@@ -486,8 +486,8 @@ def test_screen_event_projector_appends_compaction_record_and_tracks_baseline_re
 
 
 def test_screen_event_projector_does_not_reset_baseline_without_compaction_eviction() -> None:
+    from loushang.coding.presentation.tui.screen import ScreenCodingEventProjector
     from loushang.coding.ui.screen_app import ScreenCodingTuiApp
-    from loushang.coding.ui.screen_events import ScreenCodingEventProjector
 
     app = ScreenCodingTuiApp(
         model_label="kimi",
